Refactored Quest imaging tools
Separated logic that grabs/builds Quest components from the logic that invokes
the flashall.sh script.  This is useful if you want to grab/build the Quest
components and then flash on a windows machine.

Also added a --reboot-cmd option to flashall.sh that allows the Edison board to
be rebooted during the flash procedure with the specified command.